refactor: split github api and utils
